Updated packages and site headers
The 'unsafe-inline' rule was added for using critical css as it
created a inline style to help with fast loading time, as critical css
was removed the rule in not needed anymore and by removing it can
improve the general safety score in sites like HTTP Observatory from
Mozilla and security headers.

The 'access-control-allow-origin' is added to provide extra safety as it
limits who can send requests.