Updated scss to remove two deprecated methods
First the "@import" rule is being deprecated by sass, instead they
recommend "@use".

The changes in 404.scss, home.scss, page.scss and post.scss fix this
deprecation and also makes the style satisfy the
"load-partial-extension" rule in stylelint.

https://sass-lang.com/documentation/at-rules/import/

https://github.com/stylelint-scss/stylelint-scss/tree/master/src/rules/load-partial-extension

And in the header.scss and footer.scss files, existed mixed simple rules
between nested declarations, which is now a deprecated method.

As Sass will change it's behaviour to match the CSS future version, the
rules in those files are now reordered to follow this new standard.

https://sass-lang.com/documentation/breaking-changes/mixed-decls/